+++ windows/launcher/CodexDreamSkin.Launcher/GlobalUsings.cs
@@ -4,3 +4,5 @@ global using System.IO;
 global using System.Linq;
 global using System.Threading;
 global using System.Threading.Tasks;
+global using Application = System.Windows.Application;
+global using MessageBox = System.Windows.MessageBox;